fix(typography): created directory with stories of typography tokens
